FU-55 CR issues
