+++ furms-core/src/test/java/io/imunity/furms/core/resource_credits/ResourceCreditServiceImplTest.java
@@ -9,7 +9,7 @@ import io.imunity.furms.domain.resource_credits.CreateResourceCreditEvent;
 import io.imunity.furms.domain.resource_credits.RemoveResourceCreditEvent;
 import io.imunity.furms.domain.resource_credits.ResourceCredit;
 import io.imunity.furms.domain.resource_credits.UpdateResourceCreditEvent;
-import io.imunity.furms.spi.resource_credit_allocation.ResourceCreditAllocationRepository;
+import io.imunity.furms.spi.community_allocation.CommunityAllocationRepository;
 import io.imunity.furms.spi.resource_credits.ResourceCreditRepository;
 import io.imunity.furms.spi.resource_type.ResourceTypeRepository;
 import io.imunity.furms.spi.sites.SiteRepository;
@@ -38,7 +38,7 @@ class ResourceCreditServiceImplTest {
 	@Mock
 	private ResourceTypeRepository resourceTypeRepository;
 	@Mock
-	private ResourceCreditAllocationRepository resourceCreditAllocationRepository;
+	private CommunityAllocationRepository communityAllocationRepository;
 	@Mock
 	private ApplicationEventPublisher publisher;
 
@@ -48,7 +48,7 @@ class ResourceCreditServiceImplTest {
 	@BeforeEach
 	void init() {
 		MockitoAnnotations.initMocks(this);
-		ResourceCreditServiceValidator validator = new ResourceCreditServiceValidator(resourceCreditAllocationRepository, resourceCreditRepository, resourceTypeRepository, siteRepository);
+		ResourceCreditServiceValidator validator = new ResourceCreditServiceValidator(communityAllocationRepository, resourceCreditRepository, resourceTypeRepository, siteRepository);
 		service = new ResourceCreditServiceImpl(resourceCreditRepository, validator, publisher);
 		orderVerifier = inOrder(resourceCreditRepository, publisher);
 	}
