+++ furms-core/src/test/java/io/imunity/furms/core/resource_credits/ResourceCreditServiceImplValidatorTest.java
@@ -7,7 +7,7 @@ package io.imunity.furms.core.resource_credits;
 
 import io.imunity.furms.api.validation.exceptions.ResourceTypeCreditHasAllocationsRemoveValidationError;
 import io.imunity.furms.domain.resource_credits.ResourceCredit;
-import io.imunity.furms.spi.resource_credit_allocation.ResourceCreditAllocationRepository;
+import io.imunity.furms.spi.community_allocation.CommunityAllocationRepository;
 import io.imunity.furms.spi.resource_credits.ResourceCreditRepository;
 import io.imunity.furms.spi.resource_type.ResourceTypeRepository;
 import io.imunity.furms.spi.sites.SiteRepository;
@@ -35,7 +35,7 @@ class ResourceCreditServiceImplValidatorTest {
 	@Mock
 	private ResourceCreditRepository resourceCreditRepository;
 	@Mock
-	private ResourceCreditAllocationRepository resourceCreditAllocationRepository;
+	private CommunityAllocationRepository communityAllocationRepository;
 
 	@InjectMocks
 	private ResourceCreditServiceValidator validator;
@@ -263,7 +263,7 @@ class ResourceCreditServiceImplValidatorTest {
 		String id = "id";
 
 		when(resourceCreditRepository.exists(id)).thenReturn(true);
-		when(resourceCreditAllocationRepository.existsByResourceCreditId(id)).thenReturn(true);
+		when(communityAllocationRepository.existsByResourceCreditId(id)).thenReturn(true);
 
 		//when+then
 		assertThrows(ResourceTypeCreditHasAllocationsRemoveValidationError.class, () -> validator.validateDelete(id));
