+++ furms-domain/src/main/java/io/imunity/furms/domain/community_allocation/CommunityAllocationEvent.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2020 Bixbit s.c. All rights reserved.
+ * See LICENSE file for licensing information.
+ */
+
+package io.imunity.furms.domain.community_allocation;
+
+import io.imunity.furms.domain.FurmsEvent;
+
+public interface CommunityAllocationEvent extends FurmsEvent {
+}
