--- furms-domain/src/main/java/io/imunity/furms/domain/resource_credit_allocation/ResourceCreditAllocationEvent.java
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) 2020 Bixbit s.c. All rights reserved.
- * See LICENSE file for licensing information.
- */
-
-package io.imunity.furms.domain.resource_credit_allocation;
-
-import io.imunity.furms.domain.FurmsEvent;
-
-public interface ResourceCreditAllocationEvent extends FurmsEvent {
-}
