+++ furms-model-api/src/main/java/io/imunity/furms/api/community_allocation/CommunityAllocationService.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2020 Bixbit s.c. All rights reserved.
+ * See LICENSE file for licensing information.
+ */
+
+package io.imunity.furms.api.community_allocation;
+
+import io.imunity.furms.domain.community_allocation.CommunityAllocation;
+import io.imunity.furms.domain.community_allocation.CommunityAllocationResolved;
+
+import java.util.Optional;
+import java.util.Set;
+
+public interface CommunityAllocationService {
+	Optional<CommunityAllocation> findById(String id);
+
+	Optional<CommunityAllocationResolved> findByIdWithRelatedObjects(String id);
+
+	Set<CommunityAllocation> findAll();
+
+	Set<CommunityAllocationResolved> findAllWithRelatedObjects(String communityId);
+
+	void create(CommunityAllocation resourceType);
+
+	void update(CommunityAllocation resourceType);
+
+	void delete(String id);
+}
