+++ furms-server/src/main/java/io/imunity/furms/server/DemoDataInitializer.java
@@ -9,7 +9,7 @@ import io.imunity.furms.domain.communities.Community;
 import io.imunity.furms.domain.communities.CommunityGroup;
 import io.imunity.furms.domain.projects.Project;
 import io.imunity.furms.domain.projects.ProjectGroup;
-import io.imunity.furms.domain.resource_credit_allocation.ResourceCreditAllocation;
+import io.imunity.furms.domain.community_allocation.CommunityAllocation;
 import io.imunity.furms.domain.resource_credits.ResourceCredit;
 import io.imunity.furms.domain.resource_types.ResourceMeasureType;
 import io.imunity.furms.domain.resource_types.ResourceMeasureUnit;
@@ -22,7 +22,7 @@ import io.imunity.furms.spi.communites.CommunityGroupsDAO;
 import io.imunity.furms.spi.communites.CommunityRepository;
 import io.imunity.furms.spi.projects.ProjectGroupsDAO;
 import io.imunity.furms.spi.projects.ProjectRepository;
-import io.imunity.furms.spi.resource_credit_allocation.ResourceCreditAllocationRepository;
+import io.imunity.furms.spi.community_allocation.CommunityAllocationRepository;
 import io.imunity.furms.spi.resource_credits.ResourceCreditRepository;
 import io.imunity.furms.spi.resource_type.ResourceTypeRepository;
 import io.imunity.furms.spi.services.InfraServiceRepository;
@@ -54,7 +54,7 @@ class DemoDataInitializer {
 	private final InfraServiceRepository infraServiceRepository;
 	private final ResourceTypeRepository resourceTypeRepository;
 	private final ResourceCreditRepository resourceCreditRepository;
-	private final ResourceCreditAllocationRepository resourceCreditAllocationRepository;
+	private final CommunityAllocationRepository communityAllocationRepository;
 	private final SiteAgentService siteAgentService;
 
 	private String communityId;
@@ -65,7 +65,7 @@ class DemoDataInitializer {
 	                    ProjectRepository projectRepository, ProjectGroupsDAO projectGroupsDAO,
 	                    UnityServerDetector unityDetector, InfraServiceRepository infraServiceRepository,
 	                    ResourceTypeRepository resourceTypeRepository, ResourceCreditRepository resourceCreditRepository,
-	                    ResourceCreditAllocationRepository resourceCreditAllocationRepository, SiteAgentService siteAgentService) {
+	                    CommunityAllocationRepository communityAllocationRepository, SiteAgentService siteAgentService) {
 		this.communityRepository = communityRepository;
 		this.communityGroupsDAO = communityGroupsDAO;
 		this.siteRepository = siteRepository;
@@ -77,7 +77,7 @@ class DemoDataInitializer {
 		this.infraServiceRepository = infraServiceRepository;
 		this.resourceTypeRepository = resourceTypeRepository;
 		this.resourceCreditRepository = resourceCreditRepository;
-		this.resourceCreditAllocationRepository = resourceCreditAllocationRepository;
+		this.communityAllocationRepository = communityAllocationRepository;
 		this.siteAgentService = siteAgentService;
 	}
 
@@ -350,25 +350,21 @@ class DemoDataInitializer {
 			String resourceCreditBscId = resourceCreditRepository.create(resourceCreditBsc);
 			String resourceCreditBscId1 = resourceCreditRepository.create(resourceCreditBsc1);
 
-			ResourceCreditAllocation resourceCreditAllocation = ResourceCreditAllocation.builder()
-				.siteId(bscId)
+			CommunityAllocation communityAllocation = CommunityAllocation.builder()
 				.communityId(communityId)
-				.resourceTypeId(resourceTypeBscId1)
 				.resourceCreditId(resourceCreditBscId1)
 				.name("First Allocation")
 				.amount(new BigDecimal(1000))
 				.build();
-			ResourceCreditAllocation resourceCreditAllocation1 = ResourceCreditAllocation.builder()
-				.siteId(fzjId)
+			CommunityAllocation communityAllocation1 = CommunityAllocation.builder()
 				.communityId(community2Id)
-				.resourceTypeId(resourceTypeFzjId1)
 				.resourceCreditId(resourceCreditFzjId1)
 				.name("Second Allocation")
 				.amount(new BigDecimal(500))
 				.build();
 
-			resourceCreditAllocationRepository.create(resourceCreditAllocation);
-			resourceCreditAllocationRepository.create(resourceCreditAllocation1);
+			communityAllocationRepository.create(communityAllocation);
+			communityAllocationRepository.create(communityAllocation1);
 		}
 	}
 }
