--- furms-web-ui/src/main/java/io/imunity/furms/ui/views/fenix/communites/allocations/ResourceCreditAllocationFormComponent.java
@@ -1,156 +1,0 @@
-/*
- * Copyright (c) 2020 Bixbit s.c. All rights reserved.
- * See LICENSE file for licensing information.
- */
-
-package io.imunity.furms.ui.views.fenix.communites.allocations;
-
-import com.vaadin.flow.component.Composite;
-import com.vaadin.flow.component.combobox.ComboBox;
-import com.vaadin.flow.component.formlayout.FormLayout;
-import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.component.html.Label;
-import com.vaadin.flow.component.textfield.BigDecimalField;
-import com.vaadin.flow.component.textfield.TextField;
-import com.vaadin.flow.data.binder.Binder;
-import io.imunity.furms.domain.resource_types.ResourceMeasureUnit;
-import io.imunity.furms.ui.components.FurmsFormLayout;
-
-import java.math.BigDecimal;
-import java.util.Objects;
-import java.util.Optional;
-
-import static com.vaadin.flow.data.value.ValueChangeMode.EAGER;
-import static io.imunity.furms.domain.resource_types.ResourceMeasureUnit.SiUnit;
-
-public class ResourceCreditAllocationFormComponent extends Composite<Div> {
-	private static final int MAX_NAME_LENGTH = 20;
-
-	private final Binder<ResourceCreditAllocationViewModel> binder;
-
-	private ComboBox<SiteComboBoxModel> siteComboBox;
-	private ComboBox<ResourceTypeComboBoxModel> resourceTypeComboBox;
-	private ComboBox<ResourceCreditComboBoxModel> resourceCreditComboBox;
-
-	ResourceCreditAllocationFormComponent(Binder<ResourceCreditAllocationViewModel> binder, ResourceCreditAllocationComboBoxesModelsResolver resolver) {
-		this.binder = binder;
-		FormLayout formLayout = new FurmsFormLayout();
-
-		TextField nameField = new TextField();
-		nameField.setValueChangeMode(EAGER);
-		nameField.setMaxLength(MAX_NAME_LENGTH);
-		formLayout.addFormItem(nameField, getTranslation("view.fenix-admin.resource-credits-allocation.form.field.name"));
-
-		siteComboBox = new ComboBox<>();
-		siteComboBox.setItems(resolver.getSites());
-		siteComboBox.setItemLabelGenerator(site -> site.name);
-		formLayout.addFormItem(siteComboBox, getTranslation("view.fenix-admin.resource-credits-allocation.form.field.site"));
-
-		resourceTypeComboBox = new ComboBox<>();
-		resourceTypeComboBox.setItemLabelGenerator(resourceType -> resourceType.name);
-		siteComboBox.addValueChangeListener(event -> {
-			resourceTypeComboBox.setItems(resolver.getResourceTypes(event.getValue().id));
-		});
-		formLayout.addFormItem(resourceTypeComboBox, getTranslation("view.fenix-admin.resource-credits-allocation.form.field.resource_type"));
-
-		resourceCreditComboBox = new ComboBox<>();
-		resourceCreditComboBox.setItemLabelGenerator(resourceType -> resourceType.name);
-		resourceTypeComboBox.addValueChangeListener(event -> {
-			resourceCreditComboBox.setItems(resolver.getResourceCredits(event.getValue().id));
-		});
-		formLayout.addFormItem(resourceCreditComboBox, getTranslation("view.fenix-admin.resource-credits-allocation.form.field.resource_credit"));
-
-		BigDecimalField amountField = new BigDecimalField();
-		amountField.setValueChangeMode(EAGER);
-		resourceTypeComboBox.addValueChangeListener(event -> createUnitLabel(amountField, event.getValue().unit));
-		formLayout.addFormItem(amountField, getTranslation("view.fenix-admin.resource-credits-allocation.form.field.amount"));
-
-		Label availableAmountLabel = new Label();
-		resourceCreditComboBox.addValueChangeListener(event ->
-			availableAmountLabel.setText(getTranslation("view.fenix-admin.resource-credits-allocation.form.label.available") + event.getValue().amount)
-		);
-		formLayout.addFormItem(availableAmountLabel, "");
-
-		prepareValidator(nameField, siteComboBox, resourceTypeComboBox, resourceCreditComboBox, amountField);
-
-		getContent().add(formLayout);
-	}
-
-	private void createUnitLabel(BigDecimalField amountField, ResourceMeasureUnit unit) {
-		if(unit.equals(SiUnit.none))
-			return;
-		amountField.setSuffixComponent(new Label(unit.name()));
-	}
-
-	private void prepareValidator(TextField nameField, ComboBox<SiteComboBoxModel> siteComboBox,
-	                              ComboBox<ResourceTypeComboBoxModel> resourceTypeComboBox,
-	                              ComboBox<ResourceCreditComboBoxModel> resourceCreditComboBox, BigDecimalField amountField) {
-		binder.forField(nameField)
-			.withValidator(
-				value -> Objects.nonNull(value) && !value.isBlank(),
-				getTranslation("view.fenix-admin.resource-credits-allocation.form.error.validation.field.name")
-			)
-			.bind(ResourceCreditAllocationViewModel::getName, ResourceCreditAllocationViewModel::setName);
-		binder.forField(siteComboBox)
-			.withValidator(
-				Objects::nonNull,
-				getTranslation("view.fenix-admin.resource-credits-allocation.form.error.validation.combo-box.site")
-			)
-			.bind(
-				ResourceCreditAllocationViewModel::getSite,
-				ResourceCreditAllocationViewModel::setSite
-			);
-		binder.forField(resourceTypeComboBox)
-			.withValidator(
-				Objects::nonNull,
-				getTranslation("view.fenix-admin.resource-credits-allocation.form.error.validation.combo-box.resource-type")
-			)
-			.bind(
-				ResourceCreditAllocationViewModel::getResourceType,
-				ResourceCreditAllocationViewModel::setResourceType
-			);
-		binder.forField(resourceCreditComboBox)
-			.withValidator(
-				Objects::nonNull,
-				getTranslation("view.fenix-admin.resource-credits-allocation.form.error.validation.combo-box.resource-credit")
-			)
-			.bind(
-				ResourceCreditAllocationViewModel::getResourceCredit,
-				ResourceCreditAllocationViewModel::setResourceCredit
-			);
-		binder.forField(amountField)
-			.withValidator(
-				obj -> Objects.nonNull(obj) && isAmountCorrect(resourceCreditComboBox, obj),
-				getTranslation("view.fenix-admin.resource-credits-allocation.form.error.validation.field.amount")
-			)
-			.bind(
-				ResourceCreditAllocationViewModel::getAmount,
-				ResourceCreditAllocationViewModel::setAmount
-			);
-	}
-
-	private boolean isAmountCorrect(ComboBox<ResourceCreditComboBoxModel> resourceCreditComboBox, BigDecimal current) {
-		Optional<ResourceCreditComboBoxModel> value = Optional.ofNullable(resourceCreditComboBox.getValue());
-		if(value.isEmpty())
-			return false;
-		if(!value.get().split)
-			return value.get().amount.compareTo(current) == 0;
-		return value.get().amount.compareTo(current) >= 0;
-	}
-
-	public void setFormPools(ResourceCreditAllocationViewModel model) {
-		if(model.site != null) {
-			siteComboBox.setEnabled(false);
-			siteComboBox.setItems(model.site);
-		}
-		if(model.resourceType != null){
-			resourceTypeComboBox.setEnabled(false);
-			resourceTypeComboBox.setItems(model.resourceType);
-		}
-		if(model.resourceCredit != null) {
-			resourceCreditComboBox.setEnabled(false);
-			resourceCreditComboBox.setItems(model.resourceCredit);
-		}
-		binder.setBean(model);
-	}
-}
